requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-tailwind==1.0.3
Django==5.0.7
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-recaptcha==4.0.0
djangorestframework==3.15.2
idna==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
webencodings==0.5.1
